Handle remotely fetched token scopes correctly (#34)
This change removes the parsing of remotely fetched token scopes. Prior to this change it falsely assumed that Persona would return a string of scopes as `scope: "scope1 scope2`. This is now not the case (based on talis/persona-server#158). After this change and the Persona change, no matter how the token is obtained it will always eventually have `scopes: ["scope1", "scope2"]`

This flow is only triggered when encountering a token that has too many scopes to put into the payload. This will trigger the remote fetching.

Requires talis/persona-server#158